flask-vm-server/requirements.txt
Flask==1.1.1
h5py==2.9.0
numpy==1.16.5
opencv-python==4.1.0.25
Pillow==6.2.1
requests==2.22.0
gunicorn==19.9.0
google-cloud-storage